+++ config/bootstrap.php
@@ -32,6 +32,7 @@ Configure::write('WRITABLE_DIRS', array_merge(Configure::read('WRITABLE_DIRS', [
     PHOTOS,
     THUMBER_TARGET,
     UPLOADED,
+    UPLOADED . '.trash',
     USER_PICTURES,
 ]));
 
@@ -41,13 +42,6 @@ Configure::write('VENDOR_LINKS', array_merge(Configure::read('VENDOR_LINKS', [])
     'sunhater' . DS . 'kcfinder' => 'kcfinder',
     'enyo' . DS . 'dropzone' . DS . 'dist' => 'dropzone',
 ]));
-
-foreach (Configure::read('WRITABLE_DIRS') as $dir) {
-    @mkdir($dir, 0777, true);
-    if (!is_writeable($dir)) {
-        trigger_error(sprintf('Directory %s not writeable', $dir), E_USER_ERROR);
-    }
-}
 
 //Sets configuration for the Tokens plugin
 Configure::write('Tokens.usersClassOptions', [
@@ -106,5 +100,6 @@ if (!Log::getConfig('users')) {
 }
 
 require_once __DIR__ . DS . 'i18n_constants.php';
+require_once __DIR__ . DS . 'requirements.php';
 
 Type::map('jsonEntity', JsonEntityType::class);