+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # 2.x branch
+## 2.29 branch
+### 2.29.0
+* `KcFinder` has been completely replaced with `ElFinder` and all its classes
+    have been removed. `\MeCms\Command\Install\FixElFinderCommand` and
+    `\MeCms\Utility\Checkups\ElFinder` classes have been added;
+* uses and suggests `npm-asset/fancyapps-fancybox` [github](https://github.com/fancyapps/fancybox)
+    instead of `newerton/fancy-box`.
+* admin "checkup" function and all its classes have been removed.
+
 ## 2.28 branch
 ### 2.28.1
 * no longer loads the cache configuration from the application. To set custom